requirements.txt
subprocess
os
load_dotenv
logging
timedelta
TextTranslationClient
AzureKeyCredential
deepspeech
numpy
wave
whisper
python-dotenv
requests
